Fix return type of PyDict_ResolveLazyImports

+++ Include/dictobject.h
@@ -37,7 +37,7 @@ PyAPI_FUNC(int) PyDict_Contains(PyObject *mp, PyObject *key);
 PyAPI_FUNC(int) PyDict_IsLazyImport(PyObject *mp, PyObject *name);
 
 /* Forces resolving all lazy objects in a dictionary */
-PyAPI_FUNC(int) PyDict_ResolveLazyImports(PyObject *);
+PyAPI_FUNC(Py_ssize_t) PyDict_ResolveLazyImports(PyObject *);
 
 
 /* PyDict_Update(mp, other) is equivalent to PyDict_Merge(mp, other, 1). */
